Landing page updated

+++ pages/api/deleteRecipe.ts
@@ -3,6 +3,8 @@ import type {NextApiRequest, NextApiResponse} from 'next'
 import {deleteDoc, doc, setDoc} from "@firebase/firestore";
 import {db} from "../../firebase";
 import {randomUUID} from "crypto";
+import {getServerSession} from "next-auth";
+import {authOptions} from "./auth/[...nextauth]";
 
 type Data = {
     recipeId: string
@@ -22,6 +24,12 @@ export default async function handler(
         return
     }
 
+    // Check auth signed in
+    const sessionAuth = await getServerSession(req, res, authOptions)
+    if (!sessionAuth)
+        return res.status(401).send({ body: 'Unauthorised' })
+
+    // Do update
     const recipeId = req.body.recipeId as string
     await deleteDoc(
         doc(db, 'recipes', recipeId),
